Fix attempt of slow cursor when no gamepad is connected
Note: I'm not a C++ dev. My fix is, without doubt, not the good fix but it works haha

When launching "FsInputEmitter.exe -g", if there is no gamepad to find: the cursor get very slow to control, and it's hard to exit the command window. It happens only in the Mouse + Gamepad case.

I've replaced the "return false" by "return EXIT_FAILURE".
I've also changed the text color to red for getting attention to the user of the issue.